Admin: Time interval attribute now renders as `DecimalField`
Time interval field had `IntegerField` as a formfield type. This was not
intented as the value was actually considered as decimal.

Refs SHOOP-1550

+++ shoop/core/models/attributes.py
@@ -143,7 +143,7 @@ class Attribute(TranslatableModel):
         elif self.type == AttributeType.TIMEDELTA:
             kwargs.setdefault("help_text", "(as seconds)")
             # TODO: This should be more user friendly
-            return forms.IntegerField(**kwargs)
+            return forms.DecimalField(**kwargs)
         elif self.type == AttributeType.DATETIME:
             return forms.DateTimeField(**kwargs)
         elif self.type == AttributeType.DATE:
